+++ LottieViewer/MainPage.xaml.cs
@@ -6,7 +6,6 @@
 
 //#define DebugDragDrop
 using System;
-using System.Collections;
 using System.Collections.ObjectModel;
 using System.ComponentModel;
 using System.Diagnostics;
@@ -18,17 +17,14 @@ using Windows.ApplicationModel.DataTransfer;
 using Windows.Foundation.Metadata;
 using Windows.Storage;
 using Windows.Storage.Pickers;
-using Windows.System;
 using Windows.UI;
-using Windows.UI.Core;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
 using Windows.UI.Xaml.Controls.Primitives;
-using Windows.UI.Xaml.Data;
+using Windows.UI.Xaml.Documents;
 using Windows.UI.Xaml.Input;
 
 #pragma warning disable CS1591 // Missing XML comment for publicly visible type or member
-#pragma warning disable SA1402 // File may only contain a single type
 
 namespace LottieViewer
 {
@@ -127,7 +123,7 @@ namespace LottieViewer
 
                     list.Add(new PairOfStrings("Size", viewModel.SizeText));
                     list.Add(new PairOfStrings("Duration", viewModel.DurationText));
-                    list.Add(new PairOfStrings("Frames", viewModel.FrameCountText));
+                    list.Add(new PairOfStrings("Frames", $"{viewModel.FrameCountText} @ {viewModel.FramesPerSecond:0.#}fps"));
 
                     if (viewModel.Markers.Count > 0)
                     {
@@ -425,7 +421,7 @@ namespace LottieViewer
         }
 
         // Called when the user clicks on a marker hyperlink.
-        void MarkerClick(Windows.UI.Xaml.Documents.Hyperlink sender, Windows.UI.Xaml.Documents.HyperlinkClickEventArgs args)
+        void MarkerClick(Hyperlink sender, HyperlinkClickEventArgs args)
         {
             var dataContext = ((FrameworkElement)sender.ElementStart.Parent).DataContext;
             var marker = (Marker)dataContext;
@@ -433,11 +429,10 @@ namespace LottieViewer
         }
 
         // Called when the user clicks on a marker-with-duration hyperlink.
-        void MarkerEndClick(Windows.UI.Xaml.Documents.Hyperlink sender, Windows.UI.Xaml.Documents.HyperlinkClickEventArgs args)
+        void MarkerEndClick(Hyperlink sender, HyperlinkClickEventArgs args)
         {
             var dataContext = ((FrameworkElement)sender.ElementStart.Parent).DataContext;
             var marker = (MarkerWithDuration)dataContext;
-
             SeekToProgressValue(marker.ConstrainedOutProgress);
         }
 
@@ -452,83 +447,13 @@ namespace LottieViewer
             _scrubber.Focus(FocusState.Programmatic);
             _scrubber.Value = progress;
         }
-    }
 
-    // Converts bool and null values into Visibility values.
-    public sealed class VisiblityConverter : IValueConverter
-    {
-        object IValueConverter.Convert(object value, Type targetType, object parameter, string language)
+        void Page_PointerPressed(object sender, PointerRoutedEventArgs e)
         {
-            if (value is bool boolValue)
-            {
-                // The value is already a boolean.
-            }
-            else if (value is int count)
-            {
-                boolValue = count > 0;
-            }
-            else if (value is ICollection collection)
-            {
-                boolValue = collection.Count > 0;
-            }
-            else
-            {
-                // Used !null to convert to a boolean.
-                boolValue = !(value is null);
-            }
-
-            if ((string)parameter == "not")
-            {
-                // The "not" parameter inverts the logic.
-                boolValue = !boolValue;
-            }
-
-            return boolValue ? Visibility.Visible : Visibility.Collapsed;
-        }
-
-        object IValueConverter.ConvertBack(object value, Type targetType, object parameter, string language)
-        {
-            // Only support one way binding.
-            throw new NotImplementedException();
-        }
-    }
-
-    public sealed class FloatFormatter : IValueConverter
-    {
-        object IValueConverter.Convert(object value, Type targetType, object parameter, string language)
-        {
-            return ((double)value).ToString("0.#");
-        }
-
-        object IValueConverter.ConvertBack(object value, Type targetType, object parameter, string language)
-        {
-            // Only support one way binding.
-            throw new NotImplementedException();
-        }
-    }
-
-    public sealed class PropertiesTemplateSelector : DataTemplateSelector
-    {
-        public DataTemplate? Normal { get; set; }
-
-        public DataTemplate? Marker { get; set; }
-
-        public DataTemplate? MarkerWithDuration { get; set; }
-
-        protected override DataTemplate? SelectTemplateCore(object item, DependencyObject container)
-        {
-            if (item is PairOfStrings)
-            {
-                return Normal;
-            }
-            else if (item is Marker)
-            {
-                return Marker;
-            }
-            else
-            {
-                return MarkerWithDuration;
-            }
+            // By default, when the pointer is pressed, focus on the scrubber so
+            // that the arrow keys can move the scrubber.
+            e.Handled = true;
+            _scrubber.Focus(FocusState.Pointer);
         }
     }
 }
