+++ LottieViewer/ViewModel/PairOfStrings.cs
@@ -6,7 +6,7 @@ namespace LottieViewer.ViewModel
 {
     // An non-generic alternative to Tuple<string, string>.
     // Needed because XAML doesn't support generic type names.
-    sealed class PairOfStrings
+    public sealed class PairOfStrings
     {
         internal PairOfStrings(string item1, string item2) => (Item1, Item2) = (item1, item2);
